+++ pingdom/resource_pingdom_team_test.go
@@ -20,9 +20,9 @@ func TestAccResourcePingdomTeam_basic(t *testing.T) {
 	updatedName := acctest.RandomWithPrefix("tf-acc-test")
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckPingdomTeamDestroy,
+		PreCheck:          func() { testAccPreCheck(t) },
+		ProviderFactories: testAccProviderFactories,
+		CheckDestroy:      testAccCheckPingdomTeamDestroy,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccResourcePingdomTeamConfig(name),